requirements.txt
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
albumentations==1.3.1
alembic==1.16.5
altair==4.2.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
apprise==1.9.4
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.30.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
contourpy==1.3.2
coolname==2.2.0
cryptography==45.0.6
cycler==0.12.1
Cython==3.1.3
databricks-sdk==0.64.0
dateparser==1.2.2
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
duckdb==1.3.2
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
easydict==1.13
entrypoints==0.4
exceptiongroup==1.3.0
faiss-gpu==1.7.2
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
great_expectations==1.5.10
greenlet==3.2.4
griffe==1.13.0
gto==1.8.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.13.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
insightface==0.7.3
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
mistune==3.1.4
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
pandas==2.1.4
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==6.7.0
prefect==3.4.15
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python-socks==2.7.2
pytz==2025.2
PyYAML==6.0.2
qudida==0.0.4
readchar==4.2.1
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3fs==2025.7.0
s3transfer==0.13.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
scmrepo==3.5.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simsimd==6.5.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.47.3
stringzilla==3.12.6
sympy==1.14.0
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.14
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
